Move tests of OASStringType to own module'
