+++ tests/test_spec_loader.py
@@ -6,7 +6,6 @@ import pytest
 import pytest_mock as ptm
 
 from axion import spec
-from axion.spec import exceptions
 from axion.spec import model
 
 
@@ -208,30 +207,6 @@ def test_spec_load_follow_ref(
 def test_spec_load_follow_ref_no_such_ref() -> None:
     with pytest.raises(KeyError):
         spec._follow_ref({}, '#/components/schemas/Dummy')
-
-
-@pytest.mark.parametrize('default', [1, bool, {}, []])
-def test_spec_build_oas_string_default_wrong_type(default: t.Any) -> None:
-    with pytest.raises(exceptions.OASInvalidTypeValue):
-        spec._build_oas_string({'default': default})
-
-
-@pytest.mark.parametrize('example', [1, bool, {}, []])
-def test_spec_build_oas_string_example_wrong_type(example: t.Any) -> None:
-    with pytest.raises(exceptions.OASInvalidTypeValue):
-        spec._build_oas_string({'example': example})
-
-
-@pytest.mark.parametrize(('min_length', 'max_length'), [(2, 1), (10, 1)])
-def test_spec_build_oas_string_invalid_min_max_length(
-        min_length: int,
-        max_length: int,
-) -> None:
-    with pytest.raises(exceptions.OASInvalidConstraints):
-        spec._build_oas_string({
-            'minLength': min_length,
-            'maxLength': max_length,
-        })
 
 
 def test_spec_render_complex_schema() -> None:
